refactor(scripts): dedupe smoke helpers and standardize burst output
